Add AUTO_OS_UPDATES variable to update OS and packages.

+++ init/20_apt_update.sh
@@ -3,18 +3,32 @@
 # 20_apt_update.sh
 #
 
-# Update repositories
-echo "Performing updates..."
-apt-get update 2>&1 | tee /tmp/test_update
+#
+# Check update setting
+#
+AUTO_OS_UPDATES=${AUTO_OS_UPDATES:-1}
+if [ "$AUTO_OS_UPDATES" = "1" ]; then
+	#
+	# Update repositories
+	#
+	echo "Performing OS and package updates..."
+	apt-get update 2>&1 | tee /tmp/test_update
 
-# Verify that the updates will work.
-if [ "`cat /tmp/test_update | grep 'Failed'`" = "" ]; then
-	# Perform Upgrade
-	apt-get -y upgrade -o Dpkg::Options::="--force-confold"
+	#
+	# Verify that the updates will work.
+	#
+	if [ "`cat /tmp/test_update | grep 'Failed'`" = "" ]; then
+		#
+		# Perform Upgrade
+		#
+		apt-get -y upgrade -o Dpkg::Options::="--force-confold"
 
-	# Clean + purge old/obsoleted packages
-	apt-get -y autoremove
-	apt-get -y clean
-else
-	echo "Warning: Unable to update!  Check Internet connection."
+		#
+		# Clean + purge old/obsoleted packages
+		#
+		apt-get -y autoremove
+		apt-get -y clean
+	else
+		echo "Warning: Unable to update!  Check Internet connection."
+	fi
 fi
